backend/requirements.txt
alembic==1.16.5
annotated-types==0.7.0
anyio==4.10.0
bcrypt==4.3.0
cffi==2.0.0
click==8.3.0
colorama==0.4.6
cryptography==46.0.1
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
et_xmlfile==2.0.0

greenlet==3.2.4
h11==0.16.0
httptools==0.6.4
idna==3.10

Mako==1.3.10
MarkupSafe==3.0.2

openpyxl==3.1.5

passlib==1.7.4
psycopg2-binary==2.9.10
pyasn1==0.6.1
pycparser==2.23
pydantic==2.11.9
pydantic-settings==2.10.1
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-jose==3.5.0

pytz==2025.2
PyYAML==6.0.2
rsa==4.9.1

scipy==1.16.2
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.48.0
threadpoolctl==3.6.0
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2

watchfiles==1.1.0


fastapi==0.104.1
uvicorn[standard]==0.24.0
pandas==2.1.4
numpy==1.21.4
scikit-learn==1.3.2
xgboost==2.0.2
python-multipart==0.0.6
websockets==12.0
joblib==1.3.2